Move RemoveCode.cpp
