+++ services/psibase_tests/src/RemoveCode.cpp
@@ -1,8 +1,9 @@
-#include "RemoveCode.hpp"
+#include <services/test/RemoveCode.hpp>
 
 #include <psibase/dispatch.hpp>
 
 using namespace psibase;
+using namespace TestService;
 
 void RemoveCode::removeCode(psibase::Checksum256 codeHash,
                             std::uint8_t         vmType,
