+++ services/psibase_tests/test_verify_services.cpp
@@ -3,11 +3,12 @@
 
 #include <services/system/SetCode.hpp>
 #include <services/system/VerifySig.hpp>
+#include <services/test/RemoveCode.hpp>
 #include "AbortService.hpp"
-#include "RemoveCode.hpp"
 
 using namespace psibase;
 using namespace SystemService;
+using namespace TestService;
 
 namespace psibase
 {
